Change tournament purse from int to string, add purse column to tournaments table
Store the raw formatted purse string from the BallDontLie API (e.g. "$9,100,000")
instead of parsing it into an integer. This preserves currency formatting and
simplifies both ingestion and display. Removes parsePurse/formatPurse/formatCurrency
helpers across the stack. Adds a purse column to the league tournaments data table.

+++ backend/internal/features/tournaments/service.go
@@ -920,7 +920,7 @@ func toTournament(t *ent.Tournament) Tournament {
 		result.Course = *t.Course
 	}
 	if t.Purse != nil {
-		result.Purse = float64(*t.Purse)
+		result.Purse = *t.Purse
 	}
 	if t.City != nil {
 		result.City = *t.City
